Enable Har file capture in Desktop & Mobile Web
Configured Keys to Enable Har file capture in Desktop & Mobile Web.
Provide "perfecto.harfile.enable=true" in application.properties or in as a parameter in testng.xml file.
If the test is to be executed on "Windows" or "Mac" desktop OS, then we are setting "captureHAR" capability to "true" in "PerfectoDriverListener.java"
If the test is to be executed on "Android" or "iOS" devices,  then we are calling "DeviceUtils.generateHAR()" in onTestStart. And calling "DeviceUtils.stopGenerateHAR();"